update AnimationOpts, MarkPointOpts, MarkLineOpts, MarkAreaOpts
